[FLINK-31223][test] Abstract SqlClientTestBase.
